api/requirements.txt
# Pin to Windows/Render-friendly wheels (no Rust toolchain), Py3.10+
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.3.2
xgboost==1.7.6
shap==0.44.1
matplotlib==3.8.4
joblib==1.3.2
fastapi==0.110.0
uvicorn==0.29.0
gunicorn==21.2.0
pydantic==2.6.4
python-dotenv==1.0.1
optuna==3.6.1
PyYAML==6.0.1
